requirements.txt
google-cloud-bigquery
azure-mgmt-resource>=18.0.0
azure-identity>=1.5.0
